--- packages/storefront/.gitignore
@@ -1,3 +1,0 @@
-
-public/default/node_modules
-public/juizzy/node_modules
