--- packages/storefront/README.md
@@ -1,1 +1,0 @@
-# Preview Online Store : Dummy Data
