--- packages/storefront/index.html
@@ -1,19 +1,0 @@
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Storefront Playground</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-  </head>
-  <body>
-    <div id="app" data-theme="default"></div>
-
-    <script type="module">
-      import { c } from "./src/components.js";
-      document
-        .getElementById("app")
-        .setAttribute("data-components", JSON.stringify(c));
-    </script>
-    <script type="module" src="/src/app.js"></script>
-  </body>
-</html>
