--- packages/storefront/public/README.md
@@ -1,1 +1,0 @@
-# Theme in public for testing preview function
