--- packages/storefront/public/default/README.md
@@ -1,1 +1,0 @@
-# Default Theme
