--- packages/storefront/public/default/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "default",
-  "version": "1.0.0",
-  "description": "Default Theme",
-  "main": "index.js",
-  "scripts": {
-    "build": "npx tailwindcss -i ./src/main.css -o ./main.css --minify       ",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "tailwindcss": "^3.3.3"
-  }
-}
