--- packages/storefront/public/default/s-banner-promotion.html
@@ -1,1 +1,0 @@
-<span></span>
