--- packages/storefront/public/default/s-btn.html
@@ -1,1 +1,0 @@
-<button class="btn-primary">{#label#}</button>
