--- packages/storefront/public/default/s-cart-icon.html
@@ -1,1 +1,0 @@
-<span></span>
