--- packages/storefront/public/default/s-category-filter.html
@@ -1,1 +1,0 @@
-<span></span>
