--- packages/storefront/public/default/s-logo.html
@@ -1,1 +1,0 @@
-<span></span>
