--- packages/storefront/public/default/s-navigation-menus.html
@@ -1,1 +1,0 @@
-<span></span>
