--- packages/storefront/public/default/s-search-icon.html
@@ -1,1 +1,0 @@
-<span></span>
