--- packages/storefront/public/default/s-search-popup.html
@@ -1,1 +1,0 @@
-<span></span>
