--- packages/storefront/public/default/src/main.css
@@ -1,2 +1,0 @@
-@tailwind components;
-@tailwind utilities;
