--- packages/storefront/public/default/tailwind.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  content: ["./**/*.{html,js}", "./src/**/*.{html,js}"],
-};
