--- packages/storefront/public/juizzy/.gitignore
@@ -1,1 +1,0 @@
-node_modules
