--- packages/storefront/public/juizzy/README.md
@@ -1,1 +1,0 @@
-# Juizzy Theme
