--- packages/storefront/public/juizzy/index.html
@@ -1,22 +1,0 @@
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Storefront Playground</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-    <link rel="modulepreload" crossorigin href="/components.js" />
-    <link
-      rel="modulepreload"
-      crossorigin
-      href="https://unpkg.com/@orchidui/storefront@0.0.20/dist/storefront.js"
-    />
-    <link
-      rel="stylesheet"
-      href="https://unpkg.com/@orchidui/storefront@0.0.20/dist/index.css"
-    />
-  </head>
-  <body>
-    <div id="app" data-theme="juizzy"></div>
-    <script src="https://unpkg.com/@orchidui/storefront@0.0.20/dist/index.js"></script>
-  </body>
-</html>
