--- packages/storefront/public/juizzy/s-product-card.html
@@ -1,21 +1,0 @@
-<div
-  class="border border-black"
-  :style="`background-color: ${state.styles.bg}`"
->
-  <div>
-    <img class="w-full" :src="product.image" :alt="product.name" />
-  </div>
-  <div>
-    <div class="font-bold">{#product.name#}</div>
-    <div v-if="product.has_variations" class="h-[19px]">
-      {#product.variations_count#} variants
-    </div>
-    <div class="font-bold">{#product.price#}</div>
-    <s-btn
-      class="w-full"
-      label="Add to cart"
-      @click="action.addToCart(product.id)"
-    />
-    <span>{#state?.cart#}</span>
-  </div>
-</div>
