--- packages/storefront/src/App.vue
@@ -1,15 +1,0 @@
-<template>
-  <s-headers />
-
-  <!-- Preview Dummy Products -->
-
-  <!-- <s-banners/>
- <s-products/>
- <s-products/>
- <s-footer/> -->
-
-  <!-- Other Page -->
-  <!-- <s-product-detail-page /> -->
-  <!-- <s-search-filter-page /> -->
-  <!-- <s-search-product-page /> -->
-</template>
