--- packages/storefront/vite.config.js
@@ -1,30 +1,0 @@
-import { defineConfig } from "vite";
-import vue from "@vitejs/plugin-vue";
-import { resolve } from "path";
-
-export default defineConfig({
-  // publicDir: false, // remove on development : TODO : use env
-  plugins: [vue()],
-  resolve: {
-    alias: {
-      vue: "vue/dist/vue.esm-bundler.js",
-      "@": resolve(__dirname, "./src/"),
-    },
-  },
-  build: {
-    // minify: false,
-    outDir: "dist",
-    rollupOptions: {
-      input: {
-        index: resolve(__dirname, "index.html"),
-        components: resolve(__dirname, "src/components.js"),
-        storefront: resolve(__dirname, "src/storefront.js"),
-      },
-      output: {
-        entryFileNames: `[name].js`,
-        chunkFileNames: `[name].js`,
-        assetFileNames: `[name].[ext]`,
-      },
-    },
-  },
-});
