+++ src/pages/index.astro
@@ -1,10 +1,9 @@
 ---
 import Layout from '../layouts/Layout.astro';
+import OrchidUILanding from '../components/OrchidUiLanding.vue';
 ---
-<Layout title="Orchid UI">
-	<main class="w-full flex">
-		
-	</main>
+<Layout title="Orchid UI - Vue 3 Tailwind Component Library">
+	<OrchidUILanding client:only />
 </Layout>
 <style>
 </style>
